% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_images_GUI.R
\name{analyze_images_GUI}
\alias{analyze_images_GUI}
\title{Analyze Images GUI}
\usage{
analyze_images_GUI()
}
\value{
Runs an automatic image analysis
}
\description{
Brings in data entered from the "input_data_GUI" function and automatically runs a script to analyze images
pertaining to cells from fluorescent microscopy images. This uses a Shiny user-interface, where two buttons
are available to interact with. One button "Run Image Analysis" runs the analysis script while the "Close Window"
button closes the Shiny interface.
}
