% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/border.R
\name{border}
\alias{border}
\title{Border conversion}
\usage{
border(x, y)
}
\arguments{
\item{x}{An image file}

\item{y}{Dimming variable (using the 'dim' function from EBImage)}
}
\value{
Analyzing the cells on the border of \code{x}
}
\description{
Takes images to predict cells along a predetermined border in order
to count cells along the predetermined border. This will be used to help
determine those cells to "remove" them from the count function "countCells".
}
\examples{
img0 <- list.files("data/Anabena 2021/UV",pattern="tif",full.name=T)
img1 <- lapply(img0, readTIFF)
img2 <- lapply(img1, greyscale)
img3 <- lapply(img2, mapped)
img4 <- image_convert(img3)
dims=dim(img4)
border(img4,dims)
}
