% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_density.R
\name{cell_density}
\alias{cell_density}
\title{Cell density}
\usage{
cell_density(
  x,
  FOV = 0.0726,
  images = 10,
  filtration.area = 213.8,
  volume = 1,
  total.volume = 5
)
}
\arguments{
\item{x}{count results from processed image}

\item{FOV}{field of view from camera specs in square millimeters}

\item{images}{total number of fields}

\item{filtration.area}{filtration area in square millimeters}

\item{volume}{volume filtered per field in mL}

\item{total.volume}{total sample volume in mL}
}
\value{
Calculates the density of cells within your original sample, based on the inputted data
}
\description{
Takes count results from processed images to calculate cell density based
on camera field of view settings (inputted by end-user), number of images, filter area, volumes,
and amount of cells filtered that the image is visualizing.
}
\examples{
Cell.Count <- data.frame(Image_File_Name = character(0), Cell_Count = numeric(0))
images <- list.files("C:/Users/Desktop/Test_Images/",pattern = "tif", full.name = T)
images_names <- list.files("C:/Users/Desktop/Test_Images/",pattern = "tif", full.name = F)
grey_images <- greyscale(images,contrast=2)
images_mapped <- mapped(grey_images,threshold=0.3)
imagesConverted <- image_convert(imagesMapped[[1]],w=10,h=10,offset=0.001,areathresh=50,tolerance=0.8,ext=1)
count <- countCells(imagesConverted)
Cell.Count[nrow(Cell.Count) + 1, ] <- c(imgNames[[1]], count)
Cell.Count$Cell_Count<-as.numeric(Cell.Count$Cell_Count)
cell.total <- as.numeric(Cell.Count$Cell_Count)
cell.total <- sum(cell.total)
cell_density(cell.total,FOV = 0.00084,images = 10,filtration.area = 213.8,volume = 1,total.volume = 2)

}
