% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greyscale.R
\name{greyscale}
\alias{greyscale}
\title{Grey-scale conversion}
\usage{
greyscale(x, contrast = 2)
}
\arguments{
\item{x}{an image variable}

\item{contrast}{a numerical value for contrast adjustment}
}
\value{
An image variable gets converted from RGB scale to grey scale
}
\description{
Takes images to adjust the contrast and converts them to grey-scale/8-bit.
This function is needed to use the "mapping" function to properly
convert images to data-frames.
}
\examples{
images <- list.files("C:/Users/Desktop/Test_Images/",pattern = "tif", full.name = T)
greyscale(images,contrast=2)
}
