% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_convert.R
\name{image_convert}
\alias{image_convert}
\title{Image convert}
\usage{
image_convert(x, w = 17, h = 17, offset = 0.01, areathresh = 50)
}
\arguments{
\item{x}{an image variable that has been mapped using the "mapped" function}

\item{w}{width of the moving rectangular window (see EBImage)}

\item{h}{heigth of the moving rectangular window (see EBImage)}

\item{offset}{a thresholding offset}

\item{areathresh}{a numerical value of individual pixel areas of cells - this sets a threshold for what items are removed}
}
\description{
Takes a mapped image (i.e., data matrix) and counts individual cells within an image.
This is to be used with images with a single image that contains a single
fluorescent dimension of cells.
}
\examples{
img <- list.files("data/Karlodinium",pattern="tif",full.name=T)
img1 <- lapply(img[[1]], readTIFF)
img2 <- lapply(img1, greyscale, contrast = 2)
img3 <- lapply(img2, mapped, threshold = 0.3)
img4 <- image_convert(img3,w=15,h=25, offset = 0.01,areathresh=150)
}
