% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapped.R
\name{mapped}
\alias{mapped}
\title{Matrix mapping conversion}
\usage{
mapped(x, threshold = 0.3)
}
\arguments{
\item{x}{a matrix mapping conversion function}

\item{threshold}{a numerical value of pixel intensity}
}
\description{
Takes a grey-scale/8-bit image to convert to a data matrix,
where data under the numerical pixel intensity threshold is removed.
This function is needed for the use of the "image_convert" function.
}
\examples{
img <- list.files("data/Karlodinium",pattern="tif",full.name=T)
img1 <- lapply(img[[1]], readTIFF)
img2 <- lapply(img1, greyscale, contrast = 2)
img3 <- lapply(img2, mapped, threshold = 0.3)
}
