% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_image.R
\name{multi_image}
\alias{multi_image}
\title{Multi-image conversion}
\usage{
multi_image(x, y, z)
}
\arguments{
\item{x}{image variable that contain the main cells (must be created from
"outline" function)}

\item{y}{image variable that contain nuclei or other fluorescent probe (must
be created from "outline" function)}

\item{z}{layered image created using the "rgbImage" function from EBImage}
}
\description{
Takes images that have used the outline function and layers these outlines
on top of images previously brought into the R-environment. This helps
distinguish that the cells have been correctly analyzed/counted using
"image_convert" and "nuclei_segmentation" functions.
}
\examples{
F1_lab<-outline(ImgF1)
F2_lab<-outline(ImgF2)
img=rgbImage(red=ImgF2,blue=1.5*ImgF1)
multi_image(F2_lab,F1_lab,img)
}
