% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuclei_segmentation.R
\name{nuclei_segmentation}
\alias{nuclei_segmentation}
\title{Nuclei segmentation}
\usage{
nuclei_segmentation(y, z, areathresh = 50)
}
\arguments{
\item{y}{nuclei image}

\item{z}{non-nuclei image}

\item{areathresh}{area of nuclei - can be adjusted - must be a number
(automatically selected at 50)}
}
\description{
Takes nuclei microscopy image and creates blueprint upon selecting certain
cells of another microscopy image of the same group of cells to properly
segment the cells but also remove non-cell particles
}
\examples{
uv_images <- list.files("data/Anabena 2021/UV",pattern="tif",full.name=T)
tritc_images <- list.files("data/Anabena 2021/TRITC",pattern="tif",full.name=T)
images_uv <- lapply(uv_images, readTIFF)
images_tritc <- lapply(tritc_images, readTIFF)
imagesGrey_uv <- lapply(images_uv, greyscale)
imagesGrey_tritc <- lapply(images_tritc, greyscale)
imagesMapped_uv <- lapply(imagesGrey_uv, mapped)
imagesMapped_tritc <- lapply(imagesGrey_tritc, mapped)
y<-imagesMapped_uv[[1]]
z<-imagesMapped_tritc[[1]]
nuclei_segmentation(y,z,areathresh=50)
}
