% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.R
\name{outline}
\alias{outline}
\title{Outline conversion}
\usage{
outline(x)
}
\arguments{
\item{x}{normalized image variable (see example)}
}
\description{
Takes grey-scale images to convert to an outline format for the 'propagation'
function. Needed for the "multi_image" function.
}
\examples{
image_greyF2 <- readImage(paste0(grey_pathF2,grey_filesF2[f]))
ImgF2=normalize(image_greyF2,inputRange=c(0,0.75))
F2_lab<-outline(ImgF2)
}
