% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_density.R
\name{cell_density}
\alias{cell_density}
\title{Cell density}
\usage{
cell_density(
  x,
  height = 207.5,
  width = 155.6,
  diameter = 2.5,
  images = 10,
  filtration.area = 491,
  volume = 1,
  total.volume = 100
)
}
\arguments{
\item{x}{count results from processed image}

\item{height}{height of image field of view in μm (comes from camera settings)}

\item{width}{width of image field of view in μm (comes from camera settings)}

\item{diameter}{median cell diameter in μm}

\item{images}{total number of fields}

\item{filtration.area}{filtration area in mm2}

\item{volume}{volume filtered per field in mL}

\item{total.volume}{total sample volume in mL}
}
\description{
Takes count results from processed images to calculate cell density based
on camera field of view settings (inputted by end-user) and amount of cells
filtered that the image is visualizing.
}
\examples{
img <- list.files("data/Shewanella",pattern="tif",full.name=T)
img1 <- lapply(img[[1]], readTIFF)
img2 <- lapply(img1, greyscale, contrast = 2)
img3 <- lapply(img2, mapped, threshold = 0.3)
img4 <- image_convert(img3,w=15,h=25, offset = 0.01,areathresh=150)
cell.total <- countCells(img4)
cell_density(cell.total,height=203.2,width=166.4,diameter=2.5,images=15,filtration.area=491,volume=1,total.volume=100)
}
