% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countCells.R
\name{countCells}
\alias{countCells}
\title{Image count}
\usage{
countCells(x)
}
\arguments{
\item{x}{converted image variable}
}
\value{
Counting cells within \code{x}
}
\description{
Takes information from a converted image and counts cells within the image.
This uses the "border" function to remove cells along the image border and
only count cells within the image. This creates a more accurate cell count.
}
\examples{
img <- list.files("data/Karlodinium",pattern="tif",full.name=T)
img1 <- lapply(img[[1]], readTIFF)
img2 <- lapply(img1, greyscale, contrast = 2)
img3 <- lapply(img2, mapped, threshold = 0.3)
img4 <- image_convert(img3,w=15,h=25, offset = 0.01,areathresh=150)
countCells(img4)
}
