% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countImages.R
\name{countImages}
\alias{countImages}
\title{Count Images}
\usage{
countImages(image, normalize = TRUE, removeEdgeCells = TRUE)
}
\arguments{
\item{image}{A processed image variable from the "image_convert" function}

\item{normalize}{if TRUE, normalizes the resulting color image}

\item{removeEdgeCells}{automatically removes cells from along the image
border using a true (T) or false (F) input}
}
\value{
Produces a unique color-label image. Users can use the "display" function to view the image (see EBImage for more details) as well as save the image using the "writeImage" function (see readTIFF for more details).
}
\description{
Displays converted microscopy images of distinct cell counts subsequently from the "image_convert" function.
}
\examples{
images <- list.files("C:/Users/Desktop/Test_Images/",pattern = "tif", full.name = T)
grey_images <- greyscale(images,contrast=2)
images_mapped <- mapped(grey_images,threshold=0.3)
imagesConverted <- image_convert(imagesMapped[[1]],w=10,h=10,offset=0.001,areathresh=50,tolerance=0.8,ext=1)
countImages(imagesConverted, normalize = T, removeEdgeCells = T)
}
