% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greyscale.R
\name{greyscale}
\alias{greyscale}
\title{Grey-scale conversion}
\usage{
greyscale(x, contrast = 2)
}
\arguments{
\item{x}{an image variable}

\item{contrast}{a numerical value for contrast adjustment}
}
\description{
Takes images to adjust the contrast and converts them to grey-scale/8-bit.
This function is needed to use the "mapping" function to properly
convert images to data-frames.
}
\examples{
img <- list.files("data/Karlodinium",pattern="tif",full.name=T)
img1 <- lapply(img[[1]], readTIFF)
img2 <- lapply(img1, greyscale, contrast = 2)
}
